requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
numpy==1.20.1
opencv-python==4.5.1.48
pathspec==0.8.1
Pillow==8.1.0
pynetworktables==2021.0.0
regex==2020.11.13
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
